requirements.txt
neat-python
pandas
matplotlib
scikit-learn
umap-learn
requests